// src/types/generated/TerminalEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AppError } from "./AppError";

export type TerminalEvent = { "type": "data", data: string, } | { "type": "suspended" } | { "type": "resumed" } | { "type": "exit", code: number | null, } | { "type": "error", error: AppError, };